backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
prometheus-client